backend/pyproject.toml
[project]
name = "caviardeul"
version = "2.0.0"
description = ""
authors = ["julien <git@julienc.io>"]
license = "MIT"
requires-python = ">=3.11"

dependencies = [
    "django ~= 5.0",
    "gunicorn ~= 22.0",
    "psycopg[binary] ~= 3.2.1",
    "djangorestframework ~= 3.15.2",
    "httpx ~= 0.27.0",
    "cryptography ~= 42.0.8",
    "beautifulsoup4 ~= 4.12.3",
    "django-cors-headers ~= 4.4.0",
    "redis ~= 5.0.7",
]

[project.optional-dependencies]
tests = [
    "pydantic ~= 2.8.2",
    "pytest ~= 8.2.2",
    "pytest-django ~= 4.8.0",
    "pytest-cov ~= 5.0.0",
    "pytest-socket ~= 0.7.0",
    "pytest-httpx ~= 0.30.0",
    "factory-boy ~= 3.3.0"
]

[tool.coverage.run]
include = ["caviardeul/**/*.py"]
omit = ["caviardeul/tests/*"]

[tool.pytest.ini_options]
addopts = "--disable-socket --allow-hosts=redis,postgres"

[tool.ruff.lint]
extend-select = ["I"]
